refactor: Atualiza importações e configurações de teste
- Corrige importações para usar caminhos absolutos do pacote
- Remove manipulação desnecessária de sys.path
- Adiciona método load_config() público no ConfigManager
- Remove código duplicado de configuração de testes
- Atualiza imports nos testes para usar o pacote motorsport_calendar
